pyproject.toml
[tool.poetry]
name = "date-a-scientist"
version = "0.1.3"
description = "Query dataframes, find issue with your notebook snippets as if a professional data scientist was pair coding with you"
authors = ["IMPRV Dev Team <dev@imprv.ai>"]
readme = "README.md"
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.poetry.urls]
Homepage = "https://github.com/imprv-ai/date-a-scientist/sampleproject"
Issues = "https://github.com/imprv-ai/date-a-scientist/sampleproject/issues"

[tool.poetry.dependencies]
python = "^3.10"
pandasai = "^2.2.8"
pygments = "^2.18.0"
numpy = "1.26.4"

[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "^24.3.0"}
mypy = "^1.10.1"
pandas-stubs = "^2.2.2.240603"
parameterized = "^0.9.0"
pdbpp = "^0.10.3"
pre-commit = "^3.3.3"
pypyr = "^5.8.0"
pytest = "^8.2.2"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"
pytest-randomly = "^3.12.0"
python-dotenv = "^1.0.1"
twine = "^5.1.1"
types-pygments = "^2.18.0.20240506"
flake8 = "^7.1.0"
isort = "^5.13.2"
pytest-timeout = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pypyr.vars]
src_dir = "date_a_scientist"
test_dir = "tests"
python_target_version = "py311"
max_line_length = "120"
min_coverage = "50"
coverage_dir = "coverage"
linter_ignore_errors = "D100,D101,D102,D103,D104,D105,D106,D107,D202,E203,E231,W503,N806,E712"
linter_exclude = ""
dotenv_file = ".env"
test_timeout = "300"

[tool.pypyr.shortcuts]
[tool.pypyr.shortcuts.install]
    pipeline_name = "ops/install"

[tool.pypyr.shortcuts.test]
    pipeline_name = "ops/test"

[tool.pypyr.shortcuts.format]
    pipeline_name = "ops/format"

[tool.pypyr.shortcuts.check_format]
    pipeline_name = "ops/check_format"

[tool.pypyr.shortcuts.lint]
    pipeline_name = "ops/lint"

[tool.pypyr.shortcuts.precommit]
    pipeline_name = "ops/precommit"

[tool.pypyr.shortcuts.publish]
    pipeline_name = "ops/publish"
